+++ xenon-common/src/test/java/com/vmware/xenon/services/common/TestMigrationTaskService.java
@@ -32,12 +32,14 @@ import java.util.Collections;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 
 import org.junit.After;
 import org.junit.AfterClass;
@@ -339,6 +341,59 @@ public class TestMigrationTaskService extends BasicReusableHostTestCase {
         assertEquals(initState.destinationNodeGroupReference, result.destinationNodeGroupReference);
         assertEquals(initState.sourceFactoryLink, result.sourceFactoryLink);
         assertEquals(initState.sourceNodeGroupReference, result.sourceNodeGroupReference);
+    }
+
+    @Test
+    public void failMigrationWhenNodeSelectorNotConverged() throws Throwable {
+        try {
+            // Stop the last peerNode and to make the node-selector check fail.
+            final VerificationHost peerNode = this.host.getPeerHost();
+            this.host.stopHost(peerNode);
+            final VerificationHost liveNode =
+                    this.host.getInProcessHostMap().values().stream().filter(node -> node != peerNode).findFirst().get();
+            this.host.waitFor("Node did not become unavailable!",
+                    () -> {
+                        Operation op = this.sender.sendAndWait(Operation.createGet(liveNode, ServiceUriPaths.DEFAULT_NODE_GROUP));
+                        NodeGroupState ns = op.getBody(NodeGroupState.class);
+                        return this.nodeCount - 1 == ns.nodes.values().stream()
+                                .filter(ngs -> !NodeState.isUnAvailable(ngs))
+                                .collect(Collectors.toList()).size();
+                    });
+
+            // Kick-off migration
+            MigrationTaskService.State migrationState = validMigrationState(
+                    ExampleService.FACTORY_LINK);
+            TestContext ctx = this.host.testCreate(1);
+            String[] out = new String[1];
+            Operation op = Operation.createPost(this.destinationFactoryUri)
+                    .setBody(migrationState)
+                    .setReferer(this.host.getUri())
+                    .setCompletion((o, e) -> {
+                        if (e != null) {
+                            this.host.log("Post service error: %s", Utils.toString(e));
+                            ctx.failIteration(e);
+                            return;
+                        }
+                        out[0] = o.getBody(State.class).documentSelfLink;
+                        ctx.completeIteration();
+                    });
+            getDestinationHost().send(op);
+            ctx.await();
+
+            // Wait for the migration task to fail
+            State waitForServiceCompletion = waitForServiceCompletion(out[0], getDestinationHost());
+            assertEquals(waitForServiceCompletion.taskInfo.stage, TaskStage.FAILED);
+
+            // Verify that it failed because node selector availability check
+            assertTrue(waitForServiceCompletion.taskInfo.failure.message.contains(
+                    "Failed to verify availability of all node selector paths"));
+        } finally {
+            Iterator<VerificationHost> peerIt = this.host.getInProcessHostMap().values().iterator();
+            while (peerIt.hasNext()) {
+                VerificationHost peerNode = peerIt.next();
+                this.host.stopHost(peerNode);
+            }
+        }
     }
 
     @Test
@@ -647,17 +702,11 @@ public class TestMigrationTaskService extends BasicReusableHostTestCase {
     @Test
     public void successMigrateOnlyDocumentsUpdatedAfterTime() throws Throwable {
         // create object in host
-        List<ExampleServiceState> states = createExampleDocuments(this.exampleSourceFactory, getSourceHost(),
-                this.serviceCount);
-        List<URI> uris = getFullUri(getSourceHost(), states);
+        createExampleDocuments(this.exampleSourceFactory, getSourceHost(), this.serviceCount);
 
-        long time = getSourceHost()
-                .getServiceState(EnumSet.noneOf(TestProperty.class), ExampleServiceState.class, uris)
-                .values()
-                .stream()
-                .mapToLong(d -> d.documentUpdateTimeMicros)
-                .max().orElse(0);
-        assertTrue("max upateTime should not be 0", time > 0);
+        // Save the current time after the first batch has been created and
+        // before moving towards the second batch.
+        long time = Utils.getNowMicrosUtc();
 
         List<ExampleServiceState> newStates = createExampleDocuments(this.exampleSourceFactory, getSourceHost(),
                 this.serviceCount, false);
@@ -757,6 +806,78 @@ public class TestMigrationTaskService extends BasicReusableHostTestCase {
 
         getDestinationHost().getServiceState(EnumSet.noneOf(TestProperty.class),
                 ExampleServiceState.class, uris);
+    }
+
+    @Test
+    public void failMigrationWithDocumentOwnerMismatch() throws Throwable {
+        // create example services in the source node-group and make sure that each peer
+        // is owner of at-least one example service.
+        int peersWithDocs;
+        do {
+            createExampleDocuments(this.exampleSourceFactory, getSourceHost(), this.serviceCount, false);
+            ServiceDocumentQueryResult results = this.host.getExpandedFactoryState(this.exampleSourceFactory);
+            peersWithDocs = results.documents.values().stream()
+                    .map(doc -> (Utils.fromJson(doc, ServiceDocument.class)).documentOwner)
+                    .distinct()
+                    .collect(Collectors.toList())
+                    .size();
+        } while (peersWithDocs != this.nodeCount);
+
+        try {
+            // disable peer synchronization on each source host.
+            Iterator<VerificationHost> peerIt = this.host.getInProcessHostMap().values().iterator();
+            VerificationHost peerNode = null;
+            while (peerIt.hasNext()) {
+                peerNode = peerIt.next();
+                peerNode.setPeerSynchronizationEnabled(false);
+            }
+            assertNotNull(peerNode);
+
+            // Stop the last peerNode and add a new node to the group
+            this.host.stopHost(peerNode);
+            VerificationHost newHost = VerificationHost.create(0);
+            newHost.setMaintenanceIntervalMicros(TimeUnit.MILLISECONDS.toMicros(
+                    VerificationHost.FAST_MAINT_INTERVAL_MILLIS));
+            newHost.setPeerSynchronizationEnabled(false);
+            newHost.start();
+            newHost.waitForServiceAvailable(ExampleService.FACTORY_LINK);
+            this.host.addPeerNode(newHost);
+            this.host.joinNodesAndVerifyConvergence(this.nodeCount, true);
+
+            // Kick-off migration
+            MigrationTaskService.State migrationState = validMigrationState(
+                    ExampleService.FACTORY_LINK);
+            TestContext ctx = this.host.testCreate(1);
+            String[] out = new String[1];
+            Operation op = Operation.createPost(this.destinationFactoryUri)
+                    .setBody(migrationState)
+                    .setReferer(this.host.getUri())
+                    .setCompletion((o, e) -> {
+                        if (e != null) {
+                            this.host.log("Post service error: %s", Utils.toString(e));
+                            ctx.failIteration(e);
+                            return;
+                        }
+                        out[0] = o.getBody(State.class).documentSelfLink;
+                        ctx.completeIteration();
+                    });
+            getDestinationHost().send(op);
+            ctx.await();
+
+            // Wait for the migration task to fail
+            State waitForServiceCompletion = waitForServiceCompletion(out[0], getDestinationHost());
+            assertEquals(waitForServiceCompletion.taskInfo.stage, TaskStage.FAILED);
+
+            ServiceStats stats = getStats(out[0], getDestinationHost());
+            long ownerMismatchedDocuments = (long)stats.entries.get(MigrationTaskService.STAT_NAME_OWNER_MISMATCH_COUNT).latestValue;
+            assertTrue(ownerMismatchedDocuments > 0);
+        } finally {
+            Iterator<VerificationHost> peerIt = this.host.getInProcessHostMap().values().iterator();
+            while (peerIt.hasNext()) {
+                VerificationHost peerNode = peerIt.next();
+                this.host.stopHost(peerNode);
+            }
+        }
     }
 
     @Test